find transform functions

find transform functions

Usage

check_param_transform(
  param,
  param_transform,
  param_lower_bound,
  param_upper_bound
)

Arguments

param

the matrix of parameters matching the sumstat.

param_transform

a named vector of function, that to each parameter attributes a transformation function. Each function must be one of "none" (no transformation), "log" or "logit". If "logit", upper and lower values are taken in param_lower_bound param_upper_bound arguments. If the vector is of length 1, it will be recycled to all parameters. Default to "none": no transformation on any parameter.

param_lower_bound

a named vector of lower bounds for each parameter in param. If the vector is of length 1, it will be recycled to all parameters. Default to -Inf: no lower bound.

param_upper_bound

a named vector of lower bounds for each parameter in param. If the vector is of length 1, it will be recycled to all parameters. Default to Inf: no upper bound.

Value

a list, with:

transform a vector of transform for each parameter

back_transform a vector of back transform for each parameter.