Prior Goodness of Fit test

This function performs a prior Goodness of Fit (GoF) test, that checks whether we can reject the hypothesis that the target is from the same model as the particles from the sumstat matrix.

It works by sub-sampling nb.replicate points from the reference table sumstat, and then by comparing the score of the target with the distribution of scores of the calibration points.

Usage

gfit(
  target,
  sumstat,
  nb.replicate = 100,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  norm = sd,
  ncores = 1,
  nboot = 1
)

# S3 method for class 'gfit'
print(x, ...)

# S3 method for class 'gfit'
summary(object, score = "lof", k = "max", level = 0.95, ...)

# S3 method for class 'gfitsummary'
print(x, ...)

Arguments

target

a matrix of observations, with dimension Nobs x Nstats.

sumstat

a matrix of summary statistics from the training set, with dimension Ntrain x Nstats.

nb.replicate

number of replicates to draw from the training set for the calibration of the p value.

score

the score to use for calling outliers. Can be "lof", "kNN", or both.

k

The kth-distance to be used to calculate LOFs of kNNs. k can be a vector which contains multiple k values based on which the score needs to be calculated.

k_range

A vector of min and max values for k when using the "LOF max" score (see Details).

norm

the normalization function. Default to sd.

ncores

number of cores for parallel computations.

nboot

number of bootstrap replicates (see Details). Default to 1: no bootstrap replicate.

x

an object used to select a method.

...

further arguments passed to or from other methods.

object

an object for which a summary is desired.

level

the level for the confidence interval

Value

An object of class gfit. For each score, it contains a list with:

score.calib a matrix of LOF scores for the calibration data (rows), for each possible value of k (if nboot=1).

score.test a matrix of LOF scores for the test data (rows), for each possible value of k (if nboot=1).

pval a matrix of p-values for the test data (rows), for each possible value of k (if nboot=1), or an array of such matrices (if nboot>1).

n.ref the number of particles in the reference dataset (number of rows of sumstat).

n.calib the number of particles used for calibration (equal to nb.replicate).

n.boot the number of bootstrap replicates.

k_range the min and max values for k when using the "LOF max" score.

Details

The LOF and kNN scores are computations rely on the function kNN from the dbscan package.

target can be a matrix, in which case the GoF test is performed independently for each observation vector (rows of the matrix).

When using the "LOF max" statistics, the maximum LOF score over k_range is taken for all calibration and observation points.

When nboot>1, the test is performed several times, by drawing nboot different sets of nb.replicates calibration points. Highest Density Intervals are then computed using the hdi function from package HDInterval.

References

Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+), Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000. LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104. https://doi.org/10.1145/335191.335388

See also

Examples

data(gaussian_laplace)
## 1000 simulations from the Laplace dataset
sumstat <- gaussian_laplace$dataset_laplace$sim
## 10 observations from the Gaussian dataset
target <- gaussian_laplace$dataset_gaussian$sim[1:10, ]

## Apply gfit
resgfit <- gfit(target, sumstat)
resgfit
#> Prior GoF analysis
#> Using the lof with k = max
#> Number of lines in the reference table: 900 ;
#> Number of calibration point: 100 ;
#> Range of k values in the original object: k in [2, 20] ;
#> Number of target observations: 10 ;
#> Table of estimate, lower and upper 95% confidance interval pvalues:
#>       estim        lwr       upr
#>  [1,]  0.00 0.00000000 0.0000000
#>  [2,]  0.00 0.00000000 0.0000000
#>  [3,]  0.00 0.00000000 0.0000000
#>  [4,]  0.01 0.00000000 0.0295014
#>  [5,]  0.00 0.00000000 0.0000000
#>  [6,]  0.12 0.05630871 0.1836913
#>  [7,]  0.00 0.00000000 0.0000000
#>  [8,]  0.00 0.00000000 0.0000000
#>  [9,]  0.00 0.00000000 0.0000000
#> [10,]  0.00 0.00000000 0.0000000
#> Confidence intervals are based on asymptotic standard error estimation.

## Apply gfit with bootstrap
resgfit <- gfit(target, sumstat, nboot = 10) # nboot should be larger
resgfit
#> Prior GoF analysis
#> Using the lof with k = max
#> Number of lines in the reference table: 1000 ;
#> Number of calibration point: 100 ;
#> Range of k values in the original object: k in [2, 20] ;
#> Number of target observations: 10 ;
#> Table of median, lower and upper 95% HPD pvalues on 10 bootstrap replicates:
#>       median lwr upr
#>  [1,]   0.00   0 0.0
#>  [2,]   0.00   0 0.0
#>  [3,]   0.00   0 0.0
#>  [4,]   0.00   0 0.0
#>  [5,]   0.00   0 0.0
#>  [6,]   0.05   0 0.1
#>  [7,]   0.00   0 0.0
#>  [8,]   0.00   0 0.0
#>  [9,]   0.00   0 0.0
#> [10,]   0.00   0 0.0