"Frequentist" goodness of fit function

Internal function to compute the GoF using the LOF scores. The data should already be formatted and normalized. Unless you are an advanced user, you should probably use hpgfit.

Usage

gof.fit.holdout(
  data.test.obs,
  data.test.new,
  data.param,
  data.ref,
  sim.fun,
  method = c("loclinear", "ridge", "rejection"),
  kernel = c("epanechnikov", "rectangular", "gaussian", "triangular", "biweight",
    "cosine"),
  lambda = c(1e-04, 0.001, 0.01),
  trans.fun = function(x) return(x),
  back.trans.fun = function(x) return(x),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 1/100,
  split = 0.5,
  normvec,
  score = c("lof", "kNN"),
  ncores = 1,
  nboot = 1,
  ...
)

Arguments

data.test.obs

a normalized matrix of observations, with dimension Nobs x Nstats.

data.test.new

a normalized matrix of new (replicated) observations, with dimension Nobs x Nstats.

data.param

the matrix of parameters matching the sumstat.

data.ref

a normalized matrix of reference summary statistics from the training set, with dimension Ntrain x Nstats.

sim.fun

a function that, for a matrix of parameters (size nsim x nparams), returns a simulated matrix of sumstats (size nsim x nsumstats).

method

one of "rejection" (default), "loclinear" or "ridge".

kernel

for "loclinear" and "ridge", the kernel to be used.

lambda

for "ridge", the lambda penalties to be used in the ridge regression. The median of all predictions with various lambdas will be used. Default to lambda = c(0.0001, 0.001, 0.01).

trans.fun

a transformation function, that takes a matrix of parameters, and return a transformed matrix of parameters.

back.trans.fun

a transformation function, that takes a transformed matrix of parameters, and return a matrix of parameters in the original space.

k

The kth-distance to be used to calculate LOFs of kNNs. k can be a vector which contains multiple k values based on which the score needs to be calculated.

k_range

A vector of min and max values for k when using the "LOF max" score (see Details).

eps

the proportion of data points used to approximate the posterior distribution of the reference table around each observation point. Default to 0.01.

split

proportion of the posterior to be used for calibration. Default to 0.5.

normvec

the column normalization values.

score

the score to use for calling outliers. Can be "lof", "kNN", or both.

ncores

number of cores for parallel computations.

nboot

number of bootstrap replicates (see Details). Default to 1: no bootstrap replicate.

...

further arguments to be passed to sim.fun.

Value

For each score, a list, with

pval a matrix of p-values for the test data (rows), for each possible value of k.