Holdout Frequentist Goodness of Fit Test

This function performs a post-inference holdout Goodness of Fit (GoF) test, that checks whether we can reject the hypothesis that the target is from the same model as the particles from the sumstat matrix.

It works by first approximating the posterior distribution using a localization parameter eps and one of the rejection, local linear or ridge regression methods (see documentation from abc::abc). It then draws a fraction split of the reference table sumstat for calibration, and then compares the score of the target with the distribution of scores of the calibration points.

Usage

hpgfit(
  target,
  target.replica,
  param,
  sumstat,
  sim.fun,
  method = c("rejection", "loclinear", "ridge"),
  kernel = c("epanechnikov", "rectangular", "gaussian", "triangular", "biweight",
    "cosine"),
  lambda = c(1e-04, 0.001, 0.01),
  param_transform = "none",
  param_lower_bound = -Inf,
  param_upper_bound = Inf,
  trans.fun = NULL,
  back.trans.fun = NULL,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 0.01,
  split = 0.5,
  norm = sd,
  ncores = 1,
  nboot = 1,
  ...
)

# S3 method for class 'hpgfit'
print(x, ...)

# S3 method for class 'hpgfit'
summary(object, score = "lof", k = "max", level = 0.95, ...)

# S3 method for class 'hpgfitsummary'
print(x, ...)

Arguments

target

a matrix of observations, with dimension Nobs x Nstats.

target.replica

a matrix of observations, with dimension Nobs x Nstats. Each line i of the replicate target should come from the same model as the one used for the same line i of the reference table.

param

the matrix of parameters matching the sumstat.

sumstat

a matrix of summary statistics from the training set, with dimension Ntrain x Nstats.

sim.fun

a function that, for a matrix of parameters (size nsim x nparams), returns a simulated matrix of sumstats (size nsim x nsumstats).

method

one of "rejection" (default), "loclinear" or "ridge".

kernel

for "loclinear" and "ridge", the kernel to be used.

lambda

for "ridge", the lambda penalties to be used in the ridge regression. The median of all predictions with various lambdas will be used. Default to lambda = c(0.0001, 0.001, 0.01).

param_transform

a named vector of function, that to each parameter attributes a transformation function. Each function must be one of "none" (no transformation), "log" or "logit". If "logit", upper and lower values are taken in param_lower_bound param_upper_bound arguments. If the vector is of length 1, it will be recycled to all parameters. Default to "none": no transformation on any parameter.

param_lower_bound

a named vector of lower bounds for each parameter in param. If the vector is of length 1, it will be recycled to all parameters. Default to -Inf: no lower bound.

param_upper_bound

a named vector of lower bounds for each parameter in param. If the vector is of length 1, it will be recycled to all parameters. Default to Inf: no upper bound.

trans.fun

A custom transformation function, that takes as input the whole vector of parameters. If specified, it will override the default logit transform. Both trans.fun and back.trans.fun must be specified. Default to NULL.

back.trans.fun

A custom back transformation function, that takes as input the whole vector of parameters. If specified, it will override the default logit transform. Both trans.fun and back.trans.fun must be specified. Default to NULL.

score

the score to use for calling outliers. Can be "lof", "kNN", or both.

k

The kth-distance to be used to calculate LOFs of kNNs. k can be a vector which contains multiple k values based on which the score needs to be calculated.

k_range

A vector of min and max values for k when using the "LOF max" score (see Details).

eps

the proportion of data points used to approximate the posterior distribution of the reference table around each observation point. Default to 0.01.

split

proportion of the posterior to be used for calibration. Default to 0.5.

norm

the normalization function. Default to sd.

ncores

number of cores for parallel computations.

nboot

number of bootstrap replicates (see Details). Default to 1: no bootstrap replicate.

...

further arguments to be passed to sim.fun.

x

an object used to select a method.

object

an object for which a summary is desired.

level

the level for the confidence interval

Value

An object of class hpgfit. For each score, it contains a list with:

pval a matrix of p-values for the test data (rows), for each possible value of k (if nboot=1), or an array of such matrices (if nboot>1).

n.ref the number of particles in the reference dataset (number of rows of sumstat).

n.calib the number of particles used for calibration.

eps the localization fraction.

split the fraction of data used for calibration.

n.boot the number of bootstrap replicates.

k_range the min and max values for k when using the "LOF max" score.

Details

The LOF and kNN scores are computations rely on the function kNN from the dbscan package.

target can be a matrix, in which case the GoF test is performed independently for each observation vector (rows of the matrix).

When using the "LOF max" statistics, the maximum LOF score over k_range is taken for all calibration and observation points.

When nboot>1, the test is performed several times, by drawing nboot different sets of calibration points from the (fixed) posterior. Highest Density Intervals are then computed using the hdi function from package HDInterval.

References

Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+), Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000. LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104. https://doi.org/10.1145/335191.335388

See also

Examples

data(gaussian_laplace)

## 1000 simulations from the Laplace dataset
sumstat <- gaussian_laplace$dataset_laplace$sim
## Matching parameters and simulation function
param <- gaussian_laplace$dataset_laplace$param
sim_laplace <- gaussian_laplace$dataset_laplace$sim.fun

## 10 observations from the Gaussian dataset
target <- gaussian_laplace$dataset_gaussian$sim[1:10, ]
## 10 replicate obsertvations from the Gaussian dataset
param_target <- gaussian_laplace$dataset_gaussian$param[1:10, ]
target.rep <- gaussian_laplace$dataset_gaussian$sim.fun(param_target)

## Apply hpgfit
res <- hpgfit(target, target.rep, param, sumstat,
              sim.fun = sim_laplace,
              method = "rejection", eps = 0.1) # rejection method
res
#> Holdout Posterior GoF analysis
#> Using the lof with k = max
#> Number of lines in the reference table: 1000 ;
#> Number of lines in the posterior table: 100 (eps=10%);
#> Number of calibration point: 50 (split=50%);
#> Range of k values in the original object: k in [2, 20] ;
#> Number of target observations: 10 ;
#> Table of estimate, lower and upper 95% confidance interval pvalues:
#>       estim        lwr        upr
#>  [1,]  0.02 0.00000000 0.05880531
#>  [2,]  0.02 0.00000000 0.05880531
#>  [3,]  0.04 0.00000000 0.09431612
#>  [4,]  0.00 0.00000000 0.00000000
#>  [5,]  0.10 0.01684577 0.18315423
#>  [6,]  0.00 0.00000000 0.00000000
#>  [7,]  0.00 0.00000000 0.00000000
#>  [8,]  0.02 0.00000000 0.05880531
#>  [9,]  0.20 0.08912769 0.31087231
#> [10,]  0.00 0.00000000 0.00000000
#> Confidence intervals are based on asymptotic standard error estimation.

## Apply hpgfit with bootstrap
res <- hpgfit(target, target.rep, param, sumstat,
              sim.fun = sim_laplace,
              method = "rejection", eps = 0.1,
              nboot = 10) # nboot should be larger
res
#> Holdout Posterior GoF analysis
#> Using the lof with k = max
#> Number of lines in the reference table: 1000 ;
#> Number of lines in the posterior table: 100 (eps=10%);
#> Number of calibration point: 50 (split=50%);
#> Range of k values in the original object: k in [2, 20] ;
#> Number of target observations: 10 ;
#> Table of median, lower and upper 95% HPD pvalues on 10 bootstrap replicates:
#>       median  lwr  upr
#>  [1,]   0.00 0.00 0.00
#>  [2,]   0.01 0.00 0.02
#>  [3,]   0.02 0.00 0.04
#>  [4,]   0.00 0.00 0.02
#>  [5,]   0.02 0.00 0.12
#>  [6,]   0.03 0.00 0.06
#>  [7,]   0.00 0.00 0.00
#>  [8,]   0.10 0.00 0.30
#>  [9,]   0.19 0.04 0.36
#> [10,]   0.01 0.00 0.08