LOF one new point

Compute the LOF for the new point with respect to the data set of reference for various values of k. You need to have pre-computed k_dist for points in the reference data set.

Usage

lof.internal(new, set.ref, k_all, Tab.dist = NULL, Tab.id = NULL)

Arguments

new

The vector that represents the new point

set.ref

The ensemble of points in the reference data set (row by row : set.ref[j,] represents the j_th point)

k_all

specify the values of k

Tab.dist

Tab.dist[i,k] k-distance of particles i

Tab.id

Tab.id[i,1:k] k-neighborhood of particles i

Value

vector of LOF values for the new point with the respect to the reference data set, for each k in k_all.