Posterior goodness of fit function

This function is likely to be DEPRECATED in stable versions of the package.

This function performs a prior Goodness of Fit (GoF) test, that checks whether we can reject the hypothesis that the target is from the same model as the particles from the sumstat matrix.

It works by sub-sampling nb.replicate points from a localized subset of the reference table sumstat, and then by comparing the score of the target with the distribution of scores of the calibration points.

Usage

postgfit(
  target,
  sumstat,
  sumstat.replica,
  nb.replicate = 100,
  score = c("lof", "kNN"),
  k = c(2, 3, 5, 8, 14, 20),
  k_range = range(k),
  eps = 0.01,
  norm = sd,
  ncores = 1
)

Arguments

target

a matrix of observations, with dimension Nobs x Nstats.

sumstat

a matrix of summary statistics from the training set, with dimension Ntrain x Nstats.

sumstat.replica

a matrix of replicate summary statistics from the training set, with dimension Ntrain x Nstats. Each line i of the replicate summary statistic should have been simulated with the same parameters as the one used for the same line i of the reference table.

nb.replicate

number of replicates to draw from the training set for the calibration of the p value.

score

the score to use for calling outliers. Can be "lof", "kNN", or both.

k

The kth-distance to be used to calculate LOFs of kNNs. k can be a vector which contains multiple k values based on which the score needs to be calculated.

k_range

A vector of min and max values for k when using the "LOF max" score (see Details).

eps

the proportion of data points used to approximate the posterior distribution of the reference table around each observation point. Default to 0.01.

norm

the normalization function. Default to sd.

ncores

number of cores for parallel computations.

Value

For each score, it a list with:

score.calib a matrix of LOF scores for the calibration data (rows), for each possible value of k.

score.test a matrix of LOF scores for the test data (rows), for each possible value of k.

pval a matrix of p-values for the test data (rows), for each possible value of k.

n.ref the number of particles in the reference dataset (number of rows of sumstat).

n.calib the number of particles used for calibration (equal to nb.replicate).

k_range the min and max values for k when using the "LOF max" score.

References

Le Mailloux G., Bastide, P., Marin, J-M., Estoup, A. (2024+), Goodness of Fit for Bayesian Generative Models with Applications in Population Genetics.

Markus M. Breunig, Hans-Peter Kriegel, Raymond T. Ng, and Jörg Sander. 2000. LOF: identifying density-based local outliers. SIGMOD Rec. 29, 2 (June 2000), 93–104. https://doi.org/10.1145/335191.335388

See also